refactor(utils/logging/formatters): split formatters into subpackage, add module-level re-exports for backward compatibility
- Refactored apiconfig.utils.logging.formatters into a subpackage with separate modules for DetailedFormatter and RedactingFormatter.
- All public API is re-exported via __init__.py and __init__.pyi for backward compatibility.
- Updated and fixed tests to support new structure and monkeypatching.
- All quality checks (pytest, mypy, pre-commit) pass.
